requirements.txt
Flask
Flask-HTTPAuth
python-dotenv
Flask-SQLAlchemy
dash
plotly